+++ lib/octopress-ink/plugin_asset_pipeline.rb
@@ -7,7 +7,7 @@ module Octopress
       def self.compile_css(content)
         configs = sass_converter.sass_configs
         configs[:syntax] = :scss
-        configs[:style] = :compressed if Ink.config['compress_css']
+        configs[:style] = :compressed if Octopress.config['compress_css']
 
         Sass.compile(content, configs)
       end
@@ -36,7 +36,7 @@ module Octopress
         if @sass_converter
           @sass_converter
         else
-          Ink.site.converters.each do |c|
+          Octopress.site.converters.each do |c|
             @sass_converter = c if c.kind_of?(Jekyll::Converters::Sass) 
           end
           @sass_converter
@@ -157,8 +157,8 @@ module Octopress
       def self.write_combined_javascript
         js = combine_javascripts
         unless js == ''
-          if Ink.config['compress_js']
-            settings = Jekyll::Utils.symbolize_hash_keys(Ink.config['uglifier'])
+          if Octopress.config['compress_js']
+            settings = Jekyll::Utils.symbolize_hash_keys(Octopress.config['uglifier'])
             js = Uglifier.new(settings).compile(js)
           end
           write_files(js, combined_javascript_path)